Finished auto scaling example formatting, added metadata, and moved to correct service name folder.
Fix snippet tag

Fix enable vs disable metrics snippet names
