+++ .doc_gen/metadata/auto-scaling_metadata.yaml
@@ -162,7 +162,8 @@ auto-scaling_DeleteAutoScalingGroup:
         - sdk_version: 3
           github: python/example_code/auto-scaling
           excerpts:
-            - description: Update the minimum size of an Auto Scaling group to zero, terminate all
+            - description:
+                Update the minimum size of an Auto Scaling group to zero, terminate all
                 instances in the group, and delete the group.
               snippet_tags:
                 - python.cross_service.resilient_service.AutoScaler.decl
@@ -382,6 +383,15 @@ auto-scaling_TerminateInstanceInAutoScalingGroup:
                 - cpp.example_code.autoscaling.autoscaling_client
                 - cpp.example_code.autoscaling.terminate_instance_autoscaling_group1
                 - cpp.example_code.autoscaling.terminate_instance_autoscaling_group2
+    Rust:
+      versions:
+        - sdk_version: 1
+          github: rust_dev_preview/examples/autoscaling
+          excerpts:
+            - description:
+              snippet_tags:
+                - rust.autoscaling.scenario.terminate_some_instance
+                - rust.autoscaling.scenario.get_group
   services:
     auto-scaling: {TerminateInstanceInAutoScalingGroup}
 auto-scaling_SetDesiredCapacity:
@@ -444,6 +454,14 @@ auto-scaling_SetDesiredCapacity:
                 - cpp.example_code.autoscaling.autoscaling_client
                 - cpp.example_code.autoscaling.set_desired_capacity1
                 - cpp.example_code.autoscaling.set_desired_capacity2
+    Rust:
+      versions:
+        - sdk_version: 1
+          github: rust_dev_preview/examples/autoscaling
+          excerpts:
+            - description:
+              snippet_tags:
+                - rust.autoscaling.scenario.scale_desired_capacity
   services:
     auto-scaling: {SetDesiredCapacity}
 auto-scaling_DescribeAutoScalingInstances:
@@ -506,6 +524,14 @@ auto-scaling_DescribeAutoScalingInstances:
                 - cpp.example_code.autoscaling.autoscaling_client
                 - cpp.example_code.autoscaling.describe_autoscaling_instances1
                 - cpp.example_code.autoscaling.describe_autoscaling_instances2
+    Rust:
+      versions:
+        - sdk_version: 1
+          github: rust_dev_preview/examples/autoscaling
+          excerpts:
+            - description:
+              snippet_tags:
+                - rust.autoscaling.scenario.list_instances
   services:
     auto-scaling: {DescribeAutoScalingInstances}
 auto-scaling_DescribeScalingActivities:
@@ -568,6 +594,14 @@ auto-scaling_DescribeScalingActivities:
                 - cpp.example_code.autoscaling.autoscaling_client
                 - cpp.example_code.autoscaling.describe_scaling_activities1
                 - cpp.example_code.autoscaling.describe_scaling_activities2
+    Rust:
+      versions:
+        - sdk_version: 1
+          github: rust_dev_preview/examples/autoscaling
+          excerpts:
+            - description:
+              snippet_tags:
+                - rust.autoscaling.scenario.describe_scenario
   services:
     auto-scaling: {DescribeScalingActivities}
 auto-scaling_EnableMetricsCollection:
@@ -630,6 +664,14 @@ auto-scaling_EnableMetricsCollection:
                 - cpp.example_code.autoscaling.autoscaling_client
                 - cpp.example_code.autoscaling.enable_metrics_collection1
                 - cpp.example_code.autoscaling.enable_metrics_collection2
+    Rust:
+      versions:
+        - sdk_version: 1
+          github: rust_dev_preview/examples/autoscaling
+          excerpts:
+            - description:
+              snippet_tags:
+                - rust.autoscaling.scenario.enable_metrics_collection
   services:
     auto-scaling: {EnableMetricsCollection}
 auto-scaling_DisableMetricsCollection:
@@ -692,6 +734,14 @@ auto-scaling_DisableMetricsCollection:
                 - cpp.example_code.autoscaling.autoscaling_client
                 - cpp.example_code.autoscaling.disable_metrics_collection1
                 - cpp.example_code.autoscaling.disable_metrics_collection2
+    Rust:
+      versions:
+        - sdk_version: 1
+          github: rust_dev_preview/examples/autoscaling
+          excerpts:
+            - description:
+              snippet_tags:
+                - rust.autoscaling.scenario.disable_metrics_collection
   services:
     auto-scaling: {DisableMetricsCollection}
 auto-scaling_AttachLoadBalancerTargetGroups:
@@ -792,6 +842,17 @@ auto-scaling_Scenario_GroupsAndInstances:
             - description:
               snippet_tags:
                 - cpp.example_code.autoscaling.groups_and_instances_scenario
+    Rust:
+      versions:
+        - sdk_version: 1
+          github: rust_dev_preview/examples/autoscaling
+          excerpts:
+            - description:
+              snippet_files:
+                - rust_dev_preview/examples/auto-scaling/Cargo.toml
+                - rust_dev_preview/examples/auto-scaling/src/bin/scenario.rs
+                - rust_dev_preview/examples/auto-scaling/src/lib.rs
+                - rust_dev_preview/examples/auto-scaling/src/scenario.rs
   services:
     auto-scaling:
       {
