+++ rust_dev_preview/examples/auto-scaling/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "autoscaling-code-examples"
 version = "0.1.0"
-authors = ["Doug Schwartz <dougsch@amazon.com>"]
+authors = ["Doug Schwartz <dougsch@amazon.com>", "David Souther <davidsouther+github@gmail.com>"]
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
