+++ rust_dev_preview/examples/auto-scaling/src/bin/create-autoscaling-group.rs
