+++ rust_dev_preview/examples/auto-scaling/src/bin/delete-autoscaling-group.rs
