+++ rust_dev_preview/examples/auto-scaling/src/bin/update-autoscaling-group.rs
