+++ rust_dev_preview/examples/auto-scaling/src/lib.rs
