--- rust_dev_preview/examples/autoscaling/README.md
@@ -1,78 +1,0 @@
-<!--Generated by WRITEME on 2023-06-06 21:26:09.955174 (UTC)-->
-# Auto Scaling code examples for the SDK for Rust
-
-## Overview
-
-Shows how to use the AWS SDK for Rust to work with Amazon EC2 Auto Scaling.
-
-<!--custom.overview.start-->
-<!--custom.overview.end-->
-
-*Auto Scaling automatically scales EC2 instances, either with scaling policies or with scheduled scaling.*
-
-## ⚠ Important
-
-* Running this code might result in charges to your AWS account.
-* Running the tests might result in charges to your AWS account.
-* We recommend that you grant your code least privilege. At most, grant only the minimum permissions required to perform the task. For more information, see [Grant least privilege](https://docs.aws.amazon.com/IAM/latest/UserGuide/best-practices.html#grant-least-privilege).
-* This code is not tested in every AWS Region. For more information, see [AWS Regional Services](https://aws.amazon.com/about-aws/global-infrastructure/regional-product-services).
-
-<!--custom.important.start-->
-<!--custom.important.end-->
-
-## Code examples
-
-### Prerequisites
-
-For prerequisites, see the [README](../README.md#Prerequisites) in the `rust_dev_preview` folder.
-
-
-<!--custom.prerequisites.start-->
-<!--custom.prerequisites.end-->
-
-### Single actions
-
-Code excerpts that show you how to call individual service functions.
-
-* [Create a group](src/bin/create-autoscaling-group.rs#L32) (`CreateAutoScalingGroup`)
-* [Delete a group](src/bin/delete-autoscaling-group.rs#L32) (`DeleteAutoScalingGroup`)
-* [Get information about groups](src/bin/list-autoscaling-groups.rs#L24) (`DescribeAutoScalingGroups`)
-* [Update a group](src/bin/update-autoscaling-group.rs#L32) (`UpdateAutoScalingGroup`)
-
-## Run the examples
-
-### Instructions
-
-
-<!--custom.instructions.start-->
-<!--custom.instructions.end-->
-
-
-
-### Tests
-
-⚠ Running tests might result in charges to your AWS account.
-
-
-To find instructions for running these tests, see the [README](../README.md#Tests)
-in the `rust_dev_preview` folder.
-
-
-
-<!--custom.tests.start-->
-<!--custom.tests.end-->
-
-## Additional resources
-
-* [Auto Scaling User Guide](https://docs.aws.amazon.com/autoscaling/ec2/userguide/what-is-amazon-ec2-auto-scaling.html)
-* [Auto Scaling API Reference](https://docs.aws.amazon.com/autoscaling/ec2/APIReference/Welcome.html)
-* [SDK for Rust Auto Scaling reference](https://docs.rs/aws-sdk-auto-scaling/latest/aws_sdk_auto-scaling/)
-
-<!--custom.resources.start-->
-<!--custom.resources.end-->
-
----
-
-Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-
-SPDX-License-Identifier: Apache-2.0
